+++ docs/conf.py
@@ -12,7 +12,7 @@
 # All configuration values have a default; values that are commented out
 # serve to show the default.
 
-import sys, os, pkg_resources
+import sys, os, pkg_resources, datetime
 
 # If extensions (or modules to document with autodoc) are in another directory,
 # add these directories to sys.path here. If the directory is relative to the
@@ -69,7 +69,7 @@ master_doc = 'index'
 
 # General information about the project.
 project = 'Inmanta'
-copyright = '2019 Inmanta NV'
+copyright = f'{datetime.datetime.now().year} Inmanta NV'
 
 # The version info for the project you're documenting, acts as replacement for
 # |version| and |release|, also used in various other places throughout the